Fix deploy command
The s3 deploy step was failing and it is likely due to the fact that we expected the download artifacts to extract the files into a new folder. However, it was extracting to the root directory. The fix for this is to download the artifact into a given folder (provided by the `path` value) and then use that